+++ packages/zowe-explorer-api/src/tree/ZoweTreeNode.ts
@@ -26,18 +26,6 @@ export class ZoweTreeNode extends vscode.TreeItem {
     public fullPath = "";
     public dirty = false;
     public children: IZoweTreeNode[] = [];
-    /**
-     * @deprecated Define on subclass instead
-     */
-    public binaryFiles = {};
-    /**
-     * @deprecated Define on subclass instead
-     */
-    public binary = false;
-    /**
-     * @deprecated Define on subclass instead
-     */
-    public shortLabel = "";
 
     /**
      * Creates an instance of ZoweDatasetNode